Added code for CalendarStuff.java
This file provides the supporting methods for the CountTheDays program which will calculate the number of days between two dates.  It shows the use of modularization when writing Java code, and how the Java compiler can "figure things out" on its own at "compile time".  It also provides examples of proper documentation, and uses the source file header template as specified in the "Greeter.java" template program file for use in CMSI 186, Spring 2017.